+++ test/test_core.py
@@ -1,8 +1,11 @@
 # -*- coding: utf-8 -*-
+import pytest
 import re
 
 import mock
 import tunic.core
+
+from hypothesis import assume, given
 
 
 class StrDecorator(str):
@@ -212,12 +215,59 @@ def test_split_by_line_unix():
     assert ['foo', 'bar'] == tunic.core.split_by_line('foo \n bar')
 
 
+@given(str)
+def test_split_by_line_fuzz(text):
+    lines = tunic.core.split_by_line(text)
+    assert isinstance(lines, list)
+
+
+@given(str)
+def test_split_by_line_non_blank_fuzz(text):
+    assume(text.strip())
+    lines = tunic.core.split_by_line(text)
+    assert len(lines) > 0
+
+
 def test_get_current_path():
     assert '/var/www/test/current' == tunic.core.get_current_path('/var/www/test')
 
 
+def test_get_current_path_blank():
+    with pytest.raises(ValueError):
+        tunic.core.get_current_path('')
+
+
+def test_get_current_path_none():
+    with pytest.raises(ValueError):
+        tunic.core.get_current_path(None)
+
+
+@given(str)
+def test_get_current_path_fuzz(base):
+    assume(base)
+    current = tunic.core.get_current_path(base)
+    assert current.endswith('/current')
+
+
 def test_get_releases_path():
     assert '/var/www/test/releases' == tunic.core.get_releases_path('/var/www/test')
+
+
+def test_get_releases_path_blank():
+    with pytest.raises(ValueError):
+        tunic.core.get_releases_path('')
+
+
+def test_get_releases_path_none():
+    with pytest.raises(ValueError):
+        tunic.core.get_releases_path(None)
+
+
+@given(str)
+def test_get_releases_path_fuzz(base):
+    assume(base)
+    current = tunic.core.get_releases_path(base)
+    assert current.endswith('/releases')
 
 
 def test_get_release_id_with_string_version():
@@ -228,6 +278,13 @@ def test_get_release_id_with_string_version():
 def test_get_release_id_with_numeric_version():
     assert re.match(
         '^[\d]+\-1\.3\.2$', tunic.core.get_release_id('1.3.2')) is not None
+
+
+@given(str)
+def test_get_release_id_with_version_fuzz(version):
+    assume(version is not None)
+    release_id = tunic.core.get_release_id(version)
+    assert release_id.endswith('-' + version)
 
 
 def test_get_release_id_no_version():